+++ oxe-api/routes.py
@@ -112,11 +112,11 @@ from resource.private.update_my_article import UpdateMyArticle
 from resource.private.update_my_article_content import UpdateMyArticleContent
 from resource.private.update_my_form_answer import UpdateFormAnswer
 from resource.private.update_my_user import UpdateMyUser
-from resource.public.get_article_content import GetArticleContent
+from resource.public.get_public_article_content import GetPublicArticleContent
 from resource.public.get_public_article_enums import GetArticleEnums
 from resource.public.get_public_article import GetPublicArticle
 from resource.public.get_public_articles import GetPublicArticles
-from resource.public.get_related_articles import GetRelatedArticles
+from resource.public.get_public_related_articles import GetPublicRelatedArticles
 from resource.public.get_public_analytics import GetPublicAnalytics
 from resource.public.get_public_companies import GetPublicCompanies
 from resource.public.get_public_company_enums import GetPublicCompanyEnums