+++ oxe-api/test/resource/public/test_get_public_article_enums.py
@@ -1,11 +1,11 @@
 from test.BaseCase import BaseCase
 
 
-class TestGetArticleEnums(BaseCase):
+class TestGetPublicArticleEnums(BaseCase):
 
     def test_ok(self):
 
-        response = self.application.get('/public/get_article_enums')
+        response = self.application.get('/public/get_public_article_enums')
 
         self.assertEqual(200, response.status_code)
         self.assertEqual({